# day1-Swagger API/MySwaggerApi_Fixed/Models/Employee.cs
namespace MySwaggerApi.Models
{
    public class Employee
    {
        public int Id { get; set; }
        public string? Name { get; set; }
        public string? Designation { get; set; }
    }
}